Better handling for page change
Instead of running the init() function on routeChangeStart, check against the Page Loading variable,and when it returns false, run the "page has loaded" function. The entire init function doesn't need to be run every time the page changes, it's already in the app scope. This will also allow future changes to the mission edit pages.